Move ConfigCachePass from FrameworkBundle to Config
